refactor: fix empty builID in flamecomics implementation
- Remove some unwanted logging from each scanlator source codes.
- Also fix the chapter number in custom_downloader (it was put chapterName instead of chapterNumber)

- Use int64 instead of int for now (will not be needed in a re-write)

+++ internal/cstructs/download_data/download_struct.go
@@ -8,7 +8,7 @@ type DownloadData struct {
     // Number of series scrape
     TotalSeries int `json:"total_series"`
     // Number of pages(images) throughout all the series
-    TotalImages int `json:"total_images"`
+    TotalImages int64 `json:"total_images"`
     Series []SeriesData `json:"series"`
 }
 
@@ -19,7 +19,7 @@ type SeriesData struct {
     // Number of chapters the series have
     TotalChapter int `json:"total_chapter"`
     // Number of images the chapter have
-    TotalImages int `json:"total_images"`
+    TotalImages int64 `json:"total_images"`
     Chapter []ChapterData `json:"chapter"`
 }
 
@@ -28,7 +28,7 @@ type ChapterData struct {
     ChapterURL string `json:"chapter_url"`
     ChapterName string `json:"chapter_name"`
     // Number of images the chapter have
-    TotalImages int `json:"total_images"`
+    TotalImages int64 `json:"total_images"`
     Image []ImageData `json:"image"`
 }
